border for embeded video, new value bumplimit=60, subject for bumplimited post
