+++ app/config.py
@@ -1,5 +1,5 @@
 NUMBER_OF_THREADS = 30
-BUMP_LIMIT = 300
+BUMP_LIMIT = 60
 SPAN_CLASSES = ["spoiler", "strong", "italic", "strikethrough", "underline",
                 "highlight"]  # For links see FormatLinks func. Syntax is [link][/link]
