+++ requirements.txt
@@ -6,3 +6,4 @@ MarkupSafe==0.23
 mongoengine==0.10.6
 pymongo==3.2.2
 Werkzeug==0.11.10
+Pillow==3.3.0
